Admin is not allowed to delete yourself

+++ src/controllers/dashboard.js
@@ -30,22 +30,30 @@ exports.deleteData = (req, res) => {
 
     const con = db.getCon();
 
-    //in req we got user who is signed which mean that we can in that way not alow uset to not delete yourself
+    if(req.params.id == req.user.admin_id) {
+        res.send('false');
+    }
+    else {
+        con.promise().query('DELETE FROM admin_u WHERE admin_id = ?', [req.params.id])
 
-    con.promise().query('DELETE FROM admin_u WHERE admin_id = ?', [req.params.id])
-    .then(response => {
-         
+        .then(response => {
+        
         con.promise().query('SELECT * FROM admin_u')
         .then(data => {
             
             res.render('dashboard/adminManage.ejs', {data: data[0]});
         });
-      
-    }).catch(err => {
-        throw err.message;
-    });
+    
+        }).catch(err => {
+            throw err.message;
+        });
+
+    }
 }
 
+// Next we need to think it will admin have ability to 
+// change (update) others admin information or not and 
+// change information about yourself
 exports.updateData = (req, res) => {
     const con = db.getCon();
     const data = req.body;
